[승윤 Plus int 로 바꿈]

+++ src/main/java/com/ohgiraffers/project/Plus.java
@@ -1,8 +1,8 @@
 package com.ohgiraffers.project;
 
 public class Plus {
-        public void openPlus( double num1, double num2){
-            double result = num1 + num2;
+        public void openPlus( int num1, int num2){
+            int result = num1 + num2;
             System.out.println("결과: " + result);
 
         }
